Screen : Choose From Our Suggested Meals, Section : Meals Default Image from Api is Not available, Tried to fixed it. Phase : Api Implementation, Staus : On-Going

+++ lib/features/meal_details/presentation/meal_details_screen.dart
@@ -371,7 +371,7 @@ class _MealDetailsScreenState extends State<MealDetailsScreen> {
 
                             return FoodMenarelItemTileWidget(
                               imagePath: iconPath,
-                              value: menarel.kcal ?? 0,
+                              value: (menarel.kcal ?? 0).toInt(),
                               menaralName: index == 0 ? ' Carbs' : index == 1 ? ' Protein' : ' Fat',
                             );
                           },
